+++ CHANGELOG.md
@@ -1743,3 +1743,4 @@
 - [2026-06-03 22:51] feat(i18n): integrate offline caching for i18n state
 - [2026-06-03 23:00] test(auth): add unit tests for auth business logic
 - [2026-06-03 23:26] test(i18n): implement widget integration tests for i18n
+- [2026-06-04 10:03] test(theme): implement widget integration tests for theme